Check if torrent.files is empty, just in case

+++ storage.js
@@ -108,6 +108,7 @@ module.exports = function(folder, torrent) {
 	};
 
 	that.remove = function(cb) {
+		if (!torrent.files.length) return;
 		var root = torrent.files[0].path.split(path.sep)[0];
 		rimraf(path.join(folder, root), cb || noop);
 	};
